+++ src/univers/version_range.py
@@ -12,6 +12,7 @@ from semantic_version.base import AllOf
 from semantic_version.base import AnyOf
 
 from univers import gem
+from univers import maven
 from univers import versions
 from univers.utils import remove_spaces
 from univers.version_constraint import VersionConstraint
@@ -539,7 +540,9 @@ class PypiVersionRange(VersionRange):
         if ";" in string:
             raise InvalidVersionRange(f"Unsupported PyPI environment marker: {string!r}")
 
-        unsupported_chars = ";<>!=\\/|{}()`?'\"\t\n "
+        unsupported_chars = ";\\/|{}()`?'\"\t\n "
+        string = "".join(string.split(" "))
+
         if any(c in string for c in unsupported_chars):
             raise InvalidVersionRange(
                 f"Unsupported character: {unsupported_chars!r} " f"in PyPI version: {string!r}"
@@ -590,8 +593,64 @@ class MavenVersionRange(VersionRange):
     scheme = "maven"
     version_class = versions.MavenVersion
 
+    @classmethod
+    def from_native(cls, string):
+        """
+        Return a VersionRange built from a Maven version specifier ``string``.
+        """
 
-class NugetVersionRange(VersionRange):
+        string = "".join(string.split(" "))
+
+        restrictions = maven.VersionRange(string).restrictions
+        constraints = []
+
+        for restriction in restrictions:
+            lower_bound = restriction.lower_bound
+            upper_bound = restriction.upper_bound
+            lower_inclusive = restriction.lower_bound_inclusive
+            upper_inclusive = restriction.upper_bound_inclusive
+
+            if lower_bound == upper_bound:
+                constraints.append(
+                    VersionConstraint(comparator="=", version=cls.version_class(str(lower_bound)))
+                )
+                continue
+
+            if lower_bound:
+                if lower_inclusive:
+                    comparator = ">="
+                else:
+                    comparator = ">"
+                constraints.append(
+                    VersionConstraint(
+                        comparator=comparator, version=cls.version_class(str(lower_bound))
+                    )
+                )
+
+            if upper_bound:
+                if upper_inclusive:
+                    comparator = "<="
+                else:
+                    comparator = "<"
+                constraints.append(
+                    VersionConstraint(
+                        comparator=comparator, version=cls.version_class(str(upper_bound))
+                    )
+                )
+
+        return cls(constraints=constraints)
+
+    @classmethod
+    def from_natives(cls, strings):
+        if isinstance(strings, str):
+            return cls.from_native(strings)
+        constraints = []
+        for rel in strings:
+            constraints.extend(cls.from_native(rel).constraints)
+        return cls(constraints=constraints)
+
+
+class NugetVersionRange(MavenVersionRange):
     """
     NuGet range as in:[3.10.1,4)
     """